requirement.txt
python 3.11
pyperclip-1.8.2
flask_sqlalchemy
flask_login
flask
pyjwt
